Rename BilinearCompression to Downscale and update related functions<commit message>

+++ examples/Downscale/src/wave.h
@@ -22,7 +22,7 @@ struct WaveEffect {
 
 struct DrawRasterToWaveSimulator {
     DrawRasterToWaveSimulator(WaveEffect* wave_fx) : mWaveFx(wave_fx) {}
-    void draw(const vec2<int> &pt, uint32_t index, uint8_t value) {
+    void draw(const vec2<int16_t> &pt, uint32_t index, uint8_t value) {
         float valuef = value / 255.0f;
         int xx = pt.x;
         int yy = pt.y;
